requirements.txt
uvloop
aiohttp
beautifulsoup4
python-dateutil
lxml
Pillow
playwright
psutil
pyfiglet
tortoise-orm[asyncpg]
pyjwt